requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
crispy-bootstrap4==2023.1
crispy-bootstrap5==2023.10
cymem==2.0.8
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-cors-headers==4.2.0
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-import-export==3.3.8
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.4.2
drf-spectacular==0.27.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
et-xmlfile==1.1.0
filelock==3.13.1
Flask==3.0.1
fsspec==2023.12.2
h11==0.14.0
huggingface-hub==0.20.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
langcodes==3.3.0
lxml==5.1.0
MarkupPy==1.14
MarkupSafe==2.1.3
mongoengine==0.27.0
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numpy==1.26.2
odfpy==1.4.1
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
Pillow==10.1.0
pluggy==1.4.0
preshed==3.0.9
prompt-toolkit==3.0.43
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pymongo==3.12.1
PySocks==1.7.1
pytest==7.4.4
pytest-django==4.8.0
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
safetensors==0.4.1
selenium==4.18.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
sympy==1.12
tablib==3.5.0
thinc==8.2.3
tokenizers==0.15.0
torch==2.1.2
tqdm==4.66.1
transformers==4.36.2
trio==0.24.0
trio-websocket==0.11.1
typer==0.9.4
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
vine==5.1.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.0.1
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
